# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Plinth API
  version: 0.9.0
  description: |
    Recurring billing for Nigerian SaaS — card charging, bank transfers via virtual accounts,
    proration, dunning, and recovery in one API.

    All amounts are in **kobo** (1/100th of a Naira). ₦5,000 = `500000`.
servers:
  - url: https://api.useplinth.xyz
    description: Production
  - url: http://localhost:7331
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Use `sk_test_` keys for sandbox. Use `sk_live_` keys for production.
        Obtain your key from the dashboard.
...
paths:
...
  /v1/subscriptions/{id}/change-checkout:
    post:
      operationId: change-checkout
      summary: Checkout-funded plan change
      description: >
        For an upgrade when the customer has no card on file: collect the prorated amount via a
        Nomba-hosted checkout. The plan swaps when the payment settles (subscription.upgraded).
        Lateral/downgrade changes need no payment — use /change instead.
      tags: [Subscriptions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: sub_01JABC... }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id: { type: string, example: pln_01STANDARD... }
      responses:
        "200":
          description: Checkout link for the pending change
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutLink: { type: string }
                  orderReference: { type: string }
                  customerId: { type: string }
                  subscriptionId: { type: string }
                  dueMinor: { type: string, description: Prorated amount collected now (kobo). }
                  direction: { type: string, enum: [upgrade] }
                  newPlanName: { type: string }
...